+++ src/main/java/com/ixayda/iam/scim/internal/ScimRuntimeHints.java
@@ -44,8 +44,8 @@ final class ScimRuntimeHints implements RuntimeHintsRegistrar {
 			ChangePasswordConfig.class, SortConfig.class, ETagConfig.class, PaginationConfig.class,
 			AuthenticationScheme.class, ListResponse.class, SchemaResource.class, ResourceTypeResource.class,
 			ResourceTypeResource.SchemaExtension.class, AttributeDefinition.class, Meta.class, ErrorResponse.class,
-			UserResource.class, GroupResource.class, Member.class, ScimUserCreateResource.class, Name.class, Email.class,
-			PhoneNumber.class, Address.class,
+			UserResource.class, GroupResource.class, Member.class, ScimUserCreateResource.class,
+			ScimGroupCreateResource.class, Name.class, Email.class, PhoneNumber.class, Address.class,
 			InstantMessagingAddress.class, Photo.class, Group.class, Entitlement.class, Role.class,
 			X509Certificate.class, StatusSerializer.class, StatusDeserializer.class);